requirements.txt
fastapi~=0.115.9
uvicorn~=0.34.2
openai
langchain~=0.3.25
langchain-community
pinecone-client
python-dotenv~=1.1.0
pymysql
sqlalchemy~=2.0.41

pydantic~=2.11.5
langchain-core~=0.3.62
langchain-openai~=0.3.18
pinecone~=6.0.2
numpy~=2.2.6
langchain-text-splitters~=0.3.8
langchain-experimental~=0.3.4
scikit-learn~=1.6.1